pyproject.toml
[project]
name = "retail-reconcile"
version = "0.1.0"
description = "Reusable retail multi-source reconciliation + inventory health analytics"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "pydantic>=2.0",
    "python-dateutil>=2.8",
    "rapidfuzz>=3.0",
    "anthropic>=0.40",
    "jupyter>=1.0",
]

[tool.setuptools.packages.find]
include = ["retail_reconcile*"]
